// WebApp/src/app/chart/model/chart-settings.ts
export interface ChartSettings {
    xAxisText: string;
    yAxisText: string[];
    chartTitle: string;
    fontFamilyName: string;
    fontSize: number;
    pointSize: number;
    xLineVisible: boolean;
    yLineVisible: boolean;
}

export interface StateOfCharge {
    chargeFrom: number;
    chargeTo: number;
}